requirements.txt
django
djangorestframework
django-cors-headers
pandas==2.1.4
scikit-learn
requests
beautifulsoup4